Restructured project, add async requests workers, changed pattern to pair

+++ lab6/run.sh
@@ -1,14 +1,42 @@
 #!/bin/bash
 
+cd executables
+
 ./server &
 server_pid=$!
 
-./client < in.txt &
+./client < ../in.txt &
 client_pid=$!
 
-sleep 1
+for i in {1..5};
+do
 
-kill ${client_pid}
-kill ${server_pid}
+    status=0
 
-return 0
+    if [ -n "$(ps -p ${client_pid} -o pid=)" ]; then
+        status=1
+    fi
+
+    if [ -n "$(ps -p ${server_pid} -o pid=)" ]; then
+        status=1
+    fi
+
+    if [[ ${status} == 1 ]]; then
+        sleep 1
+        echo 1
+    else
+        break
+    fi
+
+done
+
+if [ -n "$(ps -p ${client_pid} -o pid=)" ]; then
+    kill ${client_pid}
+    echo "Client terminated."
+fi
+
+if [ -n "$(ps -p ${server_pid} -o pid=)" ]; then
+    kill ${server_pid}
+    echo "Server terminated."
+fi
+
